Move depends_on out of global namespace
Currently, this is in the global namespace, both in terms of file and
where the method is defined. In general, it's better to namespace
whenever possible.

In the case of a method, we can put it into a module, and then extend it
wherever it's needed. That ends up being all the base classes, which
makes sense.
